refactor: Split out utils module to more semantically sounding modules
* Currently I keep the compatibility with existing codebase by importing
  the functions/classes moved out in the utils module.
* test: Add a test case for AsyncFileWriter that uses the binary-mode.
